pycon-app-contracts/requirements.txt
eth-brownie==1.19.3
python-dotenv==1.0.0
pytest==7.4.3
eth-account==0.9.0
